+++ demo/demo.js
@@ -84,17 +84,9 @@ document.getElementById('upload').onclick = function () {
 }
 
 // onShareUrl demo
-function onShareUrlDemo (index, url) {
-  const block = {
-    id: uuid.v4(),
-    type: 'placeholder',
-    metadata: {
-      status: `Sharing... ${url}`
-    }
-  }
-  let content = ed.getContent()
-  content.splice(index, 1, block)
-  ed.setContent(content)
+function onShareUrlDemo (share) {
+  // const {block, url} = share
+  console.log(share)
   console.log('app shares url now and calls ed.setContent() with updates')
 }